requirements.txt
torch>=2.4.0
torchvision>=0.19.0 # PyTorch Computer Vision
transformers>=4.40.0 # Huggingface Transformers
sentencepiece>=0.1.99 # SentencePiece for tokenization
pillow>=10.1.0 # Python Imaging Library
pinecone>=3.0.0 # Pinecone Vector Database
numpy>=1.24.3 
pandas>=2.0.0
accelerate>=0.28.0  # Accelerator support (Mac M1/M2 or optimized CUDA)
datasets>=2.16.1
tokenizers>=0.19.0
tqdm>=4.66.1
requests>=2.31.0
streamlit>=1.42.0
python-dotenv>=1.0.0
watchdog>=4.0.0 #  for better Streamlit performance.
sentence-transformers>=2.2.2
